Add files via upload
This is a repeat of lab 4, where we were asked to create a function that computes the factorial of a number. We were tasked with writing an assembly function (lab04b.S) in ARMv8 that, when called from main, uses another assembly function (my_mul) to compute the factorial of an input, n. This version does so by using recursion.